diff --color-words reimplementation
We used to use a custom allocator to keep track of added and removed
lines in minus/plus buffers when reading from the line oriented diff.
This patch replaces the buffers with strbuf.  These lines are split
into tokens and strbuf holds one token per line.

The tokenization function is deliberately done in a stupid way but
it is called with an emit_callback structure, so that it can later
be enhanced with customization using gitattributes and other logic.
More importantly, it is not byte-oriented anymore, and would be
easier to extend it to use with UTF-8 contents.

Signed-off-by: Junio C Hamano <gitster@pobox.com>