mktag.c: adjust verify_tag parameters
The size parameter should be a size_t since it is a string length.

There is no reason that the buffer argument should not be constant except
for the NUL termination that is performed in the first few lines of this
function. This is not necessary, as the parameter given to this function was
prepared by strbuf_read() which gives a NUL terminated string.

Signed-off-by: Brandon Casey <casey@nrlssc.navy.mil>
Signed-off-by: Junio C Hamano <gitster@pobox.com>